refactor: 하나의 클래스에 존재하였던 테스트를 분리합니다
